Enable composition for 32bit cross compilation
Composition of a 32 bit cross compiled jdk for arm requires some
changes to OpenJ9.gmk as not all of the libraries were being copied
when using --with-build-jdk, and it was also using lib/compressedrefs
as the subdirectory name.

After several hours of makefile debugging I still find this composition
step rather opaque, but by following what was there previously I now
have a working build without having to do any manual shuffling of the
libraries. I've used lib/default as the subdirectory name.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>